moved model/converter to internal/converter
