+++ cmd/query/app/http_handler.go
@@ -27,7 +27,7 @@ import (
 	"github.com/jaegertracing/jaeger/internal/storage/metricstore/disabled"
 	"github.com/jaegertracing/jaeger/internal/storage/v1/api/metricstore"
 	"github.com/jaegertracing/jaeger/internal/storage/v1/api/spanstore"
-	uiconv "github.com/jaegertracing/jaeger/model/converter/json"
+	uiconv "github.com/jaegertracing/jaeger/internal/converter/json"
 	ui "github.com/jaegertracing/jaeger/model/json"
 	"github.com/jaegertracing/jaeger/pkg/jtracer"
 	"github.com/jaegertracing/jaeger/proto-gen/api_v2/metrics"
