+++ crossdock/services/agent_test.go
@@ -14,7 +14,7 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/jaegertracing/jaeger-idl/proto-gen/api_v2"
-	p2json "github.com/jaegertracing/jaeger/model/converter/json"
+	p2json "github.com/jaegertracing/jaeger/internal/converter/json"
 )
 
 var testResponse = &api_v2.SamplingStrategyResponse{
