+++ internal/converter/json/json_span_compare_test.go
