+++ internal/converter/json/process_hashtable.go
