+++ internal/converter/thrift/empty_test.go
