+++ internal/converter/thrift/jaeger/doc.go
