+++ internal/converter/thrift/jaeger/fixtures/domain_01.json
