+++ internal/converter/thrift/jaeger/fixtures/domain_02.json
