+++ internal/converter/thrift/jaeger/fixtures/domain_03.json
