+++ internal/converter/thrift/jaeger/fixtures/thrift_batch_01.json
