+++ internal/converter/thrift/jaeger/fixtures/thrift_batch_02.json
