+++ internal/converter/thrift/jaeger/package_test.go
