+++ internal/converter/thrift/jaeger/sampling_to_domain.go
