+++ internal/sampling/http/handler_test.go
@@ -20,7 +20,7 @@ import (
 	"github.com/jaegertracing/jaeger-idl/proto-gen/api_v2"
 	"github.com/jaegertracing/jaeger/internal/metricstest"
 	tSampling092 "github.com/jaegertracing/jaeger/internal/sampling/http/thrift-0.9.2"
-	p2json "github.com/jaegertracing/jaeger/model/converter/json"
+	p2json "github.com/jaegertracing/jaeger/internal/converter/json"
 )
 
 type testServer struct {
